% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{generate_seeds}
\alias{generate_seeds}
\title{Generate seeds for streams of "L'Ecuyer-CMRG" random numbers}
\usage{
generate_seeds(n)
}
\arguments{
\item{n}{Number of streams needed.}
}
\value{
A list of seeds.
}
\description{
Generate seeds for streams of "L'Ecuyer-CMRG" random numbers
}
