% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_eval}
\alias{nc_eval}
\title{Compute co-occurrence probabilities}
\usage{
nc_eval(
  occ_matrix,
  occ_probs,
  module_size = 2,
  min_occurrences = 0,
  min_support = 0
)
}
\arguments{
\item{occ_matrix}{The original occurrence matrix.}

\item{occ_probs}{The matrix of occurrence probabilities, as computed by
\code{\link[=nc_occ_probs]{nc_occ_probs()}}.}

\item{module_size}{The number of terms that should be tested for
co-occurrence.}

\item{min_occurrences}{Minimum number of occurrences of each term.}

\item{min_support}{Minimum number of occurrences of each module.}
}
\value{
A \code{data.frame} with one row for each valid module, and corresponding
number of co-occurrences and p-value.
}
\description{
The main NetCutter function. It generates p-values for all the co-occurring
modules.
}
\examples{
# Generate an occurrence matrix.
m <- matrix(FALSE, 3, 9, dimnames = list(paste0("ID", 1:3), paste0("gene", 1:9)))
m[1, 1:3] <- m[2, c(1:2, 4:5)] <- m[3, c(1, 6:9)] <- TRUE
# Set the seed using the "L'Ecuyer-CMRG" random number generator.
set.seed(1, "L'Ecuyer-CMRG")
# Compute the occurrence probabilities.
occ_probs <- nc_occ_probs(m, R = 20, S = 50)
# Evaluate the co-occurrences of pairs of terms and their statistical significance.
nc_eval(m, occ_probs, module_size = 2)
# Now evaluate triples; no need to recompute the occurrence probabilities.
nc_eval(m, occ_probs, module_size = 3)

}
