% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_occ_probs_simple}
\alias{nc_occ_probs_simple}
\title{Compute the occurrence probabilities}
\usage{
nc_occ_probs_simple(occ_matrix, R, S)
}
\arguments{
\item{occ_matrix}{The original co-occurrence matrix}

\item{R}{The number of randomisations to perform}

\item{S}{The number of successful edge swaps for each randomisation}
}
\description{
This is a simpler implementation used to check that the official
implementation (\code{\link[=nc_occ_probs]{nc_occ_probs()}}) works well.
}
