% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_randomize}
\alias{nc_randomize}
\title{Randomise the co-ocurrence matrix}
\usage{
nc_randomize(occ_matrix, S)
}
\arguments{
\item{occ_matrix}{The original co-occurrence matrix.}

\item{S}{The number of successful edge swaps to perform.}
}
\value{
A randomised copy of the co-occurrence matrix.
}
\description{
Apply an edge-swapping algorithm.
}
