% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_randomize_fast}
\alias{nc_randomize_fast}
\title{Randomize the occurrence matrix}
\usage{
nc_randomize_fast(occ_matrix, S)
}
\arguments{
\item{occ_matrix}{The original occurrence matrix.}

\item{S}{The number of successful edge swaps to perform.}
}
\description{
Faster implementation that samples row and column independently
}
