% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_randomize_R}
\alias{nc_randomize_R}
\title{Randomize the occurrence matrix}
\usage{
nc_randomize_R(occ_matrix, S)
}
\arguments{
\item{occ_matrix}{The original occurrence matrix.}

\item{S}{The number of successful edge swaps to perform.}
}
\description{
Old implementation in pure R, kept for testing purposes and for
reproducibility of old results.
}
