% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_randomize_simple}
\alias{nc_randomize_simple}
\title{Randomize the occurrence matrix}
\usage{
nc_randomize_simple(occ_matrix, S)
}
\arguments{
\item{occ_matrix}{The original occurrence matrix.}

\item{S}{The number of successful edge swaps to perform.}
}
\description{
This is a simpler implementation used to check that the official
implementation (\code{\link[=nc_randomize]{nc_randomize()}}) works well.
}
