% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter-package.R
\docType{package}
\name{netcutter-package}
\alias{netcutter}
\alias{netcutter-package}
\title{netcutter: Identification and Analysis of Co-Occurrence Networks}
\description{
R implementation of the NetCutter algorithm described in "H. Müller, F. Mancuso, Identification and analysis of co-occurrence networks with NetCutter. PLoS ONE 3, e3178 (2008)." The package identifies co-occurring terms in a list of containers. For example, it may be used to detect genes that co-occur across genomes.
}
\seealso{
Useful links:
\itemize{
  \item \doi{10.1371/journal.pone.0003178}
  \item Report bugs at \url{https://github.com/fmarotta/netcutter/issues}
}

}
\author{
\strong{Maintainer}: Federico Marotta \email{federico.marotta@embl.de}

Authors:
\itemize{
  \item Heiko Müller
  \item Francesco Mancuso
}

}
\keyword{internal}
