% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{safe_sample}
\alias{safe_sample}
\title{Sample one item from a vector, even when the vector has length 1}
\usage{
safe_sample(x)
}
\arguments{
\item{x}{Vector of values to sample}
}
\value{
One value from \code{x}.
}
\description{
Sample one item from a vector, even when the vector has length 1
}
\details{
When \code{x} has length 1, the sample() function thinks that we want to
sample from 1 to \code{x}. However, we deal want to sample vectors of unknown
length, and possibly of length 1, but we always want to sample among
the values of \code{x}. This function ensures that.
}
