% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcutter.R
\name{nc_define_modules}
\alias{nc_define_modules}
\title{Define co-occurrence modules}
\usage{
nc_define_modules(occ_matrix, terms_of_interest, module_size, min_occurrences)
}
\arguments{
\item{occ_matrix}{The original occurrence matrix.}

\item{terms_of_interest}{Vector of column names or indices representing the
terms that should be included in the analysis.}

\item{module_size}{The number of terms that should be tested for
co-occurrence.}

\item{min_occurrences}{Minimum number of occurrences of each term.}
}
\value{
A list of the valid modules.
}
\description{
Helper function to generate the list of co-occurrence terms grouped into
modules of a specified size.
}
